Fixes bug in type check 🐛 with recycle ♻
After an EOI error the program's expr value is not set to a new value,
but remains what it was before which had been recycled. This leads to a
bug where the type check gets what was supposed to be the previous
evaluation, and is also unsafe to inspect since many of the values have
been freed. This is easily fixed by setting expr to NULL which stops the
type check and cuts off reference to the recyled values.